Added doc comments and refactored classses into separate files.
